+++ src/renderer/selectors.ts
@@ -36,7 +36,7 @@ import {
     VariableImageLayerState,
     VariableLayerBase,
     VariableState,
-    VectorLayerState,
+    VectorLayerState, WebAPIServiceInfo,
     WorkflowStepState,
     WorkspaceState,
     WorldViewDataState
@@ -81,6 +81,7 @@ export const offlineModeSelector = (state: State): boolean => state.session.offl
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Remote API selectors
 
+export const webAPIServiceInfoSelector = (state: State): WebAPIServiceInfo | null => state.communication.webAPIServiceInfo;
 export const webAPIClientSelector = (state: State): WebAPIClient | null => state.communication.webAPIClient;
 export const webAPIServiceURLSelector = (state: State): string => state.communication.webAPIServiceURL;