+++ CHANGES.md
@@ -1,6 +1,10 @@
 ### Changes 3.1.1 (in dev)
 
-
+* Outputting more information to investigate into 
+  unintentionally closed WebSocket connections:
+  - Allow toasts to stay longer for such errors (2 minutes).
+  - Provide more details on the console. 
+  
 ### Changes 3.1.0
 
 * Adapted to Cate Web API 3.1.x which changed in an