+++ CHANGES.md
@@ -7,6 +7,9 @@
   unintentionally closed WebSocket connections:
   - Allow toasts to stay longer for such errors (2 minutes).
   - Provide more details on the console. 
+
+* No longer attempt to load in scratch workspaces at startup
+  (avoid error message).
   
 ### Changes 3.1.0